# perf(bitfield): bitfield unmarshaling and extended payload reading