Preallocated slices in unmarshalBitfield and switched io.ReadAll to io.ReadFull in Decode for Extended messages. Benchmark below:

**Before**
```
goos: darwin
goarch: arm64
pkg: github.com/james-lawrence/torrent/btprotocol
cpu: Apple M1 Pro
BenchmarkDecodeBitfield-10          1449        877659 ns/op     149.35 MB/s     5373023 B/op         43 allocs/op
PASS
ok      github.com/james-lawrence/torrent/btprotocol    2.595s

goos: darwin
goarch: arm64
pkg: github.com/james-lawrence/torrent/btprotocol
cpu: Apple M1 Pro
BenchmarkDecodeExtended-10         18334         67085 ns/op    1953.91 MB/s      686423 B/op         25 allocs/op
PASS
ok      github.com/james-lawrence/torrent/btprotocol    2.115s
```

**After**
```
goos: darwin
goarch: arm64
pkg: github.com/james-lawrence/torrent/btprotocol
cpu: Apple M1 Pro
BenchmarkDecodeBitfield-10          1567        792870 ns/op     165.32 MB/s     1179896 B/op          7 allocs/op
PASS
ok      github.com/james-lawrence/torrent/btprotocol    2.532s

goos: darwin
goarch: arm64
pkg: github.com/james-lawrence/torrent/btprotocol
cpu: Apple M1 Pro
BenchmarkDecodeExtended-10         95728         11966 ns/op    10954.15 MB/s     131129 B/op          7 allocs/op
PASS
ok      github.com/james-lawrence/torrent/btprotocol    1.493s

```